refactor: move test files, local project
