+++ test/test-files/data-project/config/project-scratch-def.json
@@ -1,0 +1,13 @@
+{
+  "orgName": "peter.hale company",
+  "edition": "Developer",
+  "features": ["EnableSetPasswordInApi", "ContactsToMultipleAccounts"],
+  "settings": {
+    "lightningExperienceSettings": {
+      "enableS1DesktopEnabled": true
+    },
+    "mobileSettings": {
+      "enableS1EncryptedStoragePref2": false
+    }
+  }
+}
